requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
pydantic==2.11.3
python-dotenv==1.1.0
torch==2.6
transformers==4.51.2
sentence-transformers==4.0.2
redis==5.2.1
pytest==8.3.5
httpx==0.28.1
python-jose[cryptography]==3.4.0
passlib==1.7.4
python-multipart==0.0.20
loguru==0.7.3
prometheus-client==0.21.1
numpy==2.2.4
aiohttp==3.11.16
openai==1.72.0
groq==0.22.0
pytest-asyncio==0.26.0
pytest-cov==6.1.1 
